Add convenience operators.
The SA decays into scalar -and- vector. There is ambiguity when both are supported as parameter; force the scalar variant since the reason for both paths existing is that we added optimization for scalar in this case. This is just convenience as the caller can always cast:  float(a.x) vs. a.x and ambiguity.