Update previous commit message and copyright year.
Passes minimum bar: 8 hours of regression. Fixes a potential race which was avoided by the timed lock, which can now be elided.

This is Single Consumer so there is only one task condition waiter. There can be Multiple Producers so we must notify_all to wake them all potentially, if they happen to wait simultaneously. The old design with timed lock bombarded notify one’s so eventually everyone got signalled— but with worse latency, so this improves the timings. QED.