macOS OpenGL specials..
Cocoa does not modify the Window frame when switching into the fullscreen mode. This causes the Window to return incorrect frame.. easiest work-around is to return the View’s frame which is always up-to-date. Since the Context has it’s own view, we simply return it— voila’. The easiest way to make this happen is to make the getWindowSize virtual so that context can have it’s own implementation. LOL.