Modify PNG tRNS chunk processing.
Accept incorrect PNG files because clients expect them to “work as intended”, where the specification says such files are malformed.

The root cause is that COLOR KEY (tRNS chunk) is redundant with color formats which have alpha channel stored in the file. It looks like some writers assume tRNS indicates a file with alpha channel which is patently untrue; the presense of alpha channel is encoded in the color type field of iHDR (image header).