Optimize SSE2 zero detector.
This shaves off ~25% off the encoding time. Parallel abs and coeff computation was a net loss with SSE2.

The codesize computation requires vector shift count, which is only available since AVX2 and is 32 bits so requires extra fiddling. This can also be implemented with select() but the result was slower so we’re not pursuing that yet.