Fix printf format.
The type is u64, so we must be PRIu64. llu and zu each worked on different platforms but not at the same time.

zu is for size_t
llu is for long long

I prefiously fixed the llu to zu, but it broke on another platform.. this is The Final Fix. :)

I wish compilers would just complain about it for all wrong types and not accept wrong ones, but C++ has weak typedefs so this is the result. The only real fix is for programmer to use the correct type. :)  (surprise reveal)